test: add tests for RuntimeDeadline=
Test the interactions between RuntimeDeadline= and RuntimeMaxSec=, and
make sure that the setting is correctly understood via systemd-run and
drop-in configuration.

While here, clean up the RuntimeMaxSec= tests a little bit. In
particular, use timeout instead of unbounded until loops, and use sleep
infinity in test units instead of writing a sleep loop.